+++ src/Components/TabBar.php
@@ -1,0 +1,22 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Contains the TabBar class.
+ *
+ * @copyright   Copyright (c) 2023 Vanilo UG
+ * @author      Attila Fulop
+ * @license     MIT
+ * @since       2023-09-14
+ *
+ */
+
+namespace Konekt\AppShell\Components;
+
+class TabBar extends BaseComponent
+{
+    public function __construct()
+    {
+    }
+}