+++ src/resources/database/migrations/2017_06_12_073443_create_appshell_permissions.php
@@ -22,7 +22,7 @@ class CreateAppshellPermissions extends Migration
         $adminRole = RoleProxy::create(['name' => 'admin']);
 
         $adminRole->givePermissionTo(
-            ResourcePermissions::createPermissionsForResource($this->resources)
+            ...ResourcePermissions::createPermissionsForResource($this->resources)->all()
         );
 
         $admins = UserProxy::where(['type' => UserType::ADMIN])->get();