+++ src/resources/database/migrations/2018_05_27_210939_create_settings_permissions.php
@@ -13,7 +13,7 @@ class CreateSettingsPermissions extends Migration
         $adminRole = RoleProxy::where(['name' => 'admin'])->first();
         if ($adminRole) {
             $adminRole->givePermissionTo(
-                ResourcePermissions::createPermissionsForResource('setting')
+                ...ResourcePermissions::createPermissionsForResource('setting')->all()
             );
         }
     }