+++ src/resources/database/migrations/2020_12_19_013824_add_invitation_permissions.php
@@ -28,7 +28,7 @@ class AddInvitationPermissions extends Migration
             }
 
             if ($adminRole) {
-                $adminRole->givePermissionTo($permissions);
+                $adminRole->givePermissionTo(...$permissions);
             }
         }
     }